Implemented PR changes

+++ back/src/main/java/com/lab/elephant/controller/NoteController.java
@@ -9,6 +9,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.util.List;
 import java.util.Optional;
 
 @CrossOrigin
@@ -37,8 +38,16 @@ public class NoteController {
 
   @GetMapping("{id}")
   public Note getNote(@PathVariable("id") long id) {
-    Optional<Note> note = noteService.getNote(id);
-    if (note.isPresent()) return note.get();
+    final String string = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    final User user = userService.getByEmail(string).get();
+    Optional<Note> noteOptional = noteService.getNote(id);
+    if (noteOptional.isPresent()) {
+      final Note note = noteOptional.get();
+      final List<User> usersWithPermissions = noteService.getUsersWithPermissions(note);
+      if (usersWithPermissions.contains(user))
+        return note;
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User cannot view this note");
+    }
     throw new ResponseStatusException(
             HttpStatus.NOT_FOUND, "Note Not Found");
   }
@@ -46,7 +55,15 @@ public class NoteController {
   @DeleteMapping(path = "/delete/{id}")
   public boolean deleteNote(@PathVariable("id") long id) {
     Optional<Note> optionalNote = noteService.getNote(id);
+    final String string = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    final User user = userService.getByEmail(string).get();
+    
     if (optionalNote.isPresent()) {
+      final Optional<User> owner = noteService.getOwner(optionalNote.get());
+      if (!owner.isPresent())
+        throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Note has no owner");
+      if (user.getUuid() != owner.get().getUuid())
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User can't delete this note");
       noteService.deleteNote(id);
       return true;
     }
